Split handler.rs to shell_handler.rs and protocol_handler.rs
